refactor: structure include files
