+++ gen/impl_break_word_data.h_blank
@@ -1,6 +1,6 @@
 // Unicode 14.0.0
 
-#include "internal_defines.h"
+#include <uni/internal/defines.h>
 
 UNI_ALGO_IMPL_NAMESPACE_BEGIN
 
@@ -12,4 +12,4 @@ new_stage2_break_word.txt};
 
 UNI_ALGO_IMPL_NAMESPACE_END
 
-#include "internal_undefs.h"
+#include <uni/internal/undefs.h>
