+++ include/uni/ext/translit/macedonian_to_latin_docs.h
@@ -10,7 +10,7 @@
 #include <unordered_map>
 #include <unordered_set>
 
-#include "../../cpp_uni_iterator.h"
+#include <uni/iterator.h>
 //#include "../../cpp_uni_norm.h"
 
 // If it fails a compiler messed up the encoding of this file
