+++ include/uni/impl/case_data_extern.h
