+++ include/uni/impl/norm_data.h
@@ -1,6 +1,6 @@
 // Unicode 14.0.0
 
-#include "internal_defines.h"
+#include <uni/internal/defines.h>
 
 UNI_ALGO_IMPL_NAMESPACE_BEGIN
 
@@ -17582,4 +17582,4 @@ uaix_const_data type_codept stage3_comp[][64] = {
 
 UNI_ALGO_IMPL_NAMESPACE_END
 
-#include "internal_undefs.h"
+#include <uni/internal/undefs.h>
