+++ include/uni/impl/norm_data_extern.h
