+++ include/uni/internal/sentinel.h
