+++ perf/perf_break_utf8.h
@@ -15,10 +15,11 @@
 #include <chrono>
 #include <vector>
 #include <stdexcept>
-#include "../src/cpp_uni_convert.h"
-#include "../src/cpp_uni_break_grapheme.h"
-#include "../src/cpp_uni_break_word.h"
-#include "../src/cpp_uni_ranges.h"
+
+#include <uni/convert.h>
+#include <uni/break_grapheme.h>
+#include <uni/break_word.h>
+#include <uni/ranges.h>
 
 int break_uni(std::string_view str)
 {
