+++ test/main.h
@@ -2,13 +2,13 @@
  * License: Public Domain or MIT - choose whatever you want.
  * See LICENSE.md */
 
-#include "../src/cpp_uni_convert.h"
-#include "../src/cpp_uni_case.h"
-#include "../src/cpp_uni_norm.h"
-#include "../src/cpp_uni_iterator.h"
-#include "../src/cpp_uni_break_grapheme.h"
-#include "../src/cpp_uni_break_word.h"
-#include "../src/cpp_uni_ranges.h"
+#include <uni/convert.h>
+#include <uni/case.h>
+#include <uni/norm.h>
+#include <uni/iterator.h>
+#include <uni/break_grapheme.h>
+#include <uni/break_word.h>
+#include <uni/ranges.h>
 
 #include <stdexcept>
 #include <iostream>
